Clean text output from url component
