+++ src/backend/base/langflow/helpers/data.py
@@ -139,3 +139,38 @@ def messages_to_text(template: str, messages: Message | list[Message]) -> str:
 
     formated_messages = [template.format(data=message.model_dump(), **message.model_dump()) for message in messages_]
     return "\n".join(formated_messages)
+
+
+def _safe_convert(self, data: Any) -> str:
+    """Safely convert input data to string."""
+    try:
+        if isinstance(data, str):
+            return data
+        if isinstance(data, Message):
+            return data.get_text()
+        if isinstance(data, Data):
+            if data.get_text() is None:
+                msg = "Empty Data object"
+                raise ValueError(msg)
+            return data.get_text()
+        if isinstance(data, DataFrame):
+            if self.clean_data:
+                # Remove empty rows
+                data = data.dropna(how="all")
+                # Remove empty lines in each cell
+                data = data.replace(r"^\s*$", "", regex=True)
+                # Replace multiple newlines with a single newline
+                data = data.replace(r"\n+", "\n", regex=True)
+
+            # Replace pipe characters to avoid markdown table issues
+            processed_data = data.replace(r"\|", r"\\|", regex=True)
+
+            processed_data = processed_data.map(
+                lambda x: str(x).replace("\n", "<br/>") if isinstance(x, str) else x
+            )
+
+            return processed_data.to_markdown(index=False)
+        return str(data)
+    except (ValueError, TypeError, AttributeError) as e:
+        msg = f"Error converting data: {e!s}"
+        raise ValueError(msg) from e
